requirements.txt
langchain
openai
streamlit
os
python-dotenv
langchain-community